Move logging out of root import
This is to avoid confusing import of system and local module
